Endpoints och statisk metod för valuta omvandling färdiga, logg och säkerhet kvar

+++ src/main/java/com/greger/wigelltravels/service/DestinationServiceImpl.java
@@ -3,6 +3,7 @@ package com.greger.wigelltravels.service;
 import com.greger.wigelltravels.dao.DestinationRepository;
 import com.greger.wigelltravels.entity.Address;
 import com.greger.wigelltravels.entity.Destination;
+import com.greger.wigelltravels.entity.Trip;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -46,9 +47,34 @@ public class DestinationServiceImpl implements DestinationService{
     }
 
     @Override
+    public Destination updateDestination(int id, Destination destination) {
+        Destination destinationFromDb = findById(id);
+        destinationFromDb.setHotellName(destination.getHotellName());
+        destinationFromDb.setPricePerWeek(destination.getPricePerWeek());
+        destinationFromDb.setCity(destination.getCity());
+        destinationFromDb.setCountry(destination.getCountry());
+        return save(destinationFromDb);
+    }
+
+    @Override
     @Transactional
-    public void deleteById(int id) {
-        destinationRepository.deleteById(id);
+    public String deleteById(int id) {
+        String message = "";
+        List<Trip> tripList = destinationRepository.findAllByDestination(findById(id).getId());
+        System.out.println("RESA BASERAT PÅÅ RESA: " + tripList);
+        System.out.println(tripList.size());
+        if (tripList.size() == 0){
+            message = "Destination with id: " + id + " has been deleted!";
+            destinationRepository.deleteById(id);
+            return message;
+        }else {
+            message = "Destination with id: " + id + " can not be deleted! Destination is used in " + tripList.size() + " trips!";
+            for (Trip t : tripList){
+                String tripId = "\n" + "Trip ID: " + String.valueOf(t.getTripId());
+                message = message.concat(tripId);
+            }
+        }
+        return message;
     }
 
     @Override
@@ -65,14 +91,12 @@ public class DestinationServiceImpl implements DestinationService{
             System.out.println("FRÅN DB: " + destinationFromDatabase);
             return destinationFromDatabase;
         }
-        destination.setId(0);
         if(autoSave){
+            destination.setId(0);
             destinationFromDatabase = save(destination);
             System.out.println("SPARAD: " + destinationFromDatabase);
             return destinationFromDatabase;
         }
-
-
         System.out.println("###############################################################################");
       return destination;
     }
